redo SaveHashLookup()
Previously I added every element of SL.Global.Hashlookup to a string and then wrote it to a file all at once. Started getting out of memory errors when writing a 1mb+ hashlookup file. Switched to adding one line at a time and it seems to be better. Needs to be tested with bigger libraries (3k+?)