fix opening song bug
experiment keeps track of what the last played song/group is since it can't rely on the music wheel. it's loaded as a preference but needs to be outside of the normal SL preference reader since it's not a whitelisted value. With this fix, starting the game should jump you straight to the last song played instead of the first song from the first group