Improve tagging
TaggedSongs.txt and Tags.txt were originally part of the download which meant when updating versions it was easy to accidentally overwrite your own one when updating the theme. Took them out so now we just create the text files if they're not there. Additionally, start with a default tag if none exist so those without keyboards can still use the tag menu right away without having to go in to the file to create some tags.