add more fields to MeasureCounterAndModsLevel.lua
previously things like percent of stream, total stream measures, and breakdowns were calculated in pane display. instead calculate them when parsing a stream and then store it in SL player table in case it needs to be used anywhere else.

This is a preparation step for eventually calculating crossovers and footswitches as well