fix group bar graph error
when a pack contains a chart with a really high difficulty the text wasn't displaying properly. cap everything at 25 now so all 25+ charts will be in one group. if those charts become more prevalent it might be a good idea to raise this to 30 or something but for now there are not that many 25+ charts in most packs so i'm leaving it alone.