update TechParser
Begin updating TechParser to figure out which arrow should be hit with which foot. Currently it only calculates for arrows that are > half the peak bpm (peak in a single measure). The reason for this is stamina songs often have easy doublesteps/crossovers in break sections that are slow to the point they don't matter (to me). I want the parser to tell me if there are crossovers and other tech in the streams (don't care about breaks). Thus, when determining what foot should hit each arrow we also only look at the faster notes