errorbar and tech update
Move errorbar to (mostly) fit in no matter where the measure counter is. Move tech to modules since it'll be called by pane display and screeneval now. Show per foot stats on screen eval even if tech isn't turned on (if it isn't, run the parser on screen eval) this should reduce lag on select music for those that don't need to know if there are crossovers and stuff. Considering making it an on demand thing instead of automatic for every song as well