Add quick options
Now you can choose to change either the rate or the scroll speed without going in to the options menu. Rate has a couple minor issues with the sample music not changing properly but i don't even know what it should be so i left it alone. For changing the scroll speed, M and C mods were simple but X is kind of weird. Right now I have it where it maintains the X mod instead of the adjusted speed. But that means if you play songs at different bpms in a row you have to change it everytime. I think this is how X players use it?? Not sure... Makes more sense to me to maintain close to the same adjusted speed but then you might as well play M so I guess it should maintain X instead.