add test input to gameplay
testing must be done to make sure it doesn't cause lag or anything. also it flashes so fast at high speeds it might be useless lol. might be cool for viewers to look at though. also considering moving it to the other side and making it bigger when center single player is on. need to see how distracting it is.