remove selectplaymode
for now at least, remove play modes to force fantasy. FA+ and marathon should both work from within fantasy and this is not really a theme for casuals. the option to choose a default is still in SL options which i should remove eventually because now it doesn't do anything